--- tests/Models/TarjamaModel.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace LaravelArab\Tarjama\Tests\Models;
-
-use LaravelArab\Tarjama\Models\Translation;
-
-class TarjamaModel extends Translation
-{
-    /**
-     * The table associated with the model.
-     *
-     * @var string
-     */
-    protected $table = 'tarjamas';
-}
