+++ tests/Models/TestsTranslationsModel.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Fevrok\Translatable\Tests\Models;
+
+use Fevrok\Translatable\Models\Translation;
+
+class TestsTranslationsModel extends Translation
+{
+    /**
+     * The table associated with the model.
+     *
+     * @var string
+     */
+    protected $table = 'tests_translations';
+}
